refactor cmcapiclient, move config to types. todo: init for cmcapiclient, return response in router
