+++ cmcapiclient/cmcapiclient.go
@@ -5,55 +5,29 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"time"
+
+	"github.com/chrobles/go-rest-api/types"
 )
 
 // Client : client for interacting with coinmarketcap API
 type Client struct {
+	BaseURL string
 	Key     string
-	Address string
+	Local   bool
 }
 
-// RangeData : resulting json from query against cm API
-type RangeData struct {
-	Status struct {
-		Timestamp    time.Time   `json:"timestamp"`
-		ErrorCode    int         `json:"error_code"`
-		ErrorMessage interface{} `json:"error_message"`
-		Elapsed      int         `json:"elapsed"`
-		CreditCount  int         `json:"credit_count"`
-	} `json:"status"`
-	Data []struct {
-		ID                int         `json:"id"`
-		Name              string      `json:"name"`
-		Symbol            string      `json:"symbol"`
-		Slug              string      `json:"slug"`
-		NumMarketPairs    int         `json:"num_market_pairs"`
-		DateAdded         time.Time   `json:"date_added"`
-		Tags              []string    `json:"tags"`
-		MaxSupply         int         `json:"max_supply"`
-		CirculatingSupply int         `json:"circulating_supply"`
-		TotalSupply       int         `json:"total_supply"`
-		Platform          interface{} `json:"platform"`
-		CmcRank           int         `json:"cmc_rank"`
-		LastUpdated       time.Time   `json:"last_updated"`
-		Quote             struct {
-			USD struct {
-				Price            float64   `json:"price"`
-				Volume24H        float64   `json:"volume_24h"`
-				PercentChange1H  float64   `json:"percent_change_1h"`
-				PercentChange24H float64   `json:"percent_change_24h"`
-				PercentChange7D  float64   `json:"percent_change_7d"`
-				MarketCap        float64   `json:"market_cap"`
-				LastUpdated      time.Time `json:"last_updated"`
-			} `json:"USD"`
-		} `json:"quote"`
-	} `json:"data"`
+// Init : apply client configuration
+func (client *Client) Init(cfg types.Config) error {
+	client.BaseURL = cfg.CoinMarketCap.BaseURL
+	client.Key = cfg.CoinMarketCap.Key
+	client.Local = cfg.CoinMarketCap.Local
+
+	return nil
 }
 
-// NewRangeRequest : generate a request for a range of data from coinmarketcap API
-func (client *Client) NewRangeRequest(start string, limit string) *http.Request {
-	req, err := http.NewRequest("GET", client.Address, nil)
+// NewMarketRequest : generate a request for a range of data from coinmarketcap API
+func (client *Client) NewMarketRequest(start string, limit string) *http.Request {
+	req, err := http.NewRequest("GET", client.BaseURL, nil)
 	if err != nil {
 		log.Print(err)
 		return nil
@@ -72,12 +46,19 @@ func (client *Client) NewRangeRequest(start string, limit string) *http.Request 
 }
 
 // Get : do a request and return the results
-func (client *Client) Get(req *http.Request) RangeData {
-	res, _ := http.DefaultClient.Do(req)
+func (client *Client) Get(req *http.Request) (*types.MarketListings, error) {
+	var (
+		res     *http.Response
+		resdata *types.MarketListings
+	)
+
+	res, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, err
+	}
 	defer res.Body.Close()
 
-	var resData RangeData
-	_ = json.NewDecoder(res.Body).Decode(&resData)
+	_ = json.NewDecoder(res.Body).Decode(resdata)
 
-	return resData
+	return resdata, nil
 }
