+++ go.mod
@@ -3,7 +3,7 @@ module github.com/chrobles/go-rest-api
 go 1.13
 
 require (
-	github.com/davecgh/go-spew v1.1.1
 	github.com/gin-gonic/gin v1.5.0
+	github.com/joho/godotenv v1.3.0
 	github.com/kelseyhightower/envconfig v1.4.0
 )
